add guard to message routes and add refresh token

+++ src/messages/message.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { Message } from './entities/message.entity';
 import { CreateMessageDto } from './dto/create-message.dto';
 import { UpdateMessageDto } from './dto/update-message.dto';
@@ -6,6 +10,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { PersonService } from 'src/person/person.service';
 import { PaginationDto } from 'src/app/common/dto/pagination.dto';
+import { TokenPayloadDto } from 'src/auth/dto/token-payload.dto';
 
 @Injectable()
 export class MessageService {
@@ -50,11 +55,12 @@ export class MessageService {
     this.throwNotFoundError();
   }
 
-  async create(createMessageDto: CreateMessageDto) {
-    const { fromId, toId } = createMessageDto;
-
-    const from = await this.personService.findOne(fromId);
-    const to = await this.personService.findOne(toId);
+  async create(
+    createMessageDto: CreateMessageDto,
+    tokenPayload: TokenPayloadDto,
+  ) {
+    const from = await this.personService.findOne(tokenPayload.sub);
+    const to = await this.personService.findOne(createMessageDto.toId);
 
     const newMessage = {
       text: createMessageDto.text,
@@ -71,15 +77,24 @@ export class MessageService {
       ...message,
       from: {
         id: message.from.id,
+        name: message.from.name,
       },
       to: {
         id: message.to.id,
+        name: message.to.name,
       },
     };
   }
 
-  async update(id: number, updateMessageDto: UpdateMessageDto) {
+  async update(
+    id: number,
+    updateMessageDto: UpdateMessageDto,
+    tokenPayload: TokenPayloadDto,
+  ) {
     const message = await this.findOne(id);
+
+    if (message.from.id !== tokenPayload.sub)
+      throw new ForbiddenException('Action Unauthorized');
 
     message.text = updateMessageDto.text ?? message.text;
     message.readed = updateMessageDto.readed ?? message.readed;
@@ -91,10 +106,13 @@ export class MessageService {
     return message;
   }
 
-  async remove(id: number) {
-    const message = await this.messageRepository.findOneBy({ id });
+  async remove(id: number, tokenPayload: TokenPayloadDto) {
+    const message = await this.findOne(id);
 
     if (!message) return this.throwNotFoundError();
+
+    if (message.from.id !== tokenPayload.sub)
+      throw new ForbiddenException('Action Unauthorized');
 
     return this.messageRepository.remove(message);
   }
